Update branch with : Make code more readeable. Revisit the locking of io_instruction insertion in stream. Optimistic heuristic to avoid locking the running thread.
Add KAAPI_PIPELINE_GPUTASK in kaapi_impl.h.
If it is defined then activate the algorithm to control the order of task insertions,
communications and kernel launches. With KAAPI_PIPELINE_GPUTASK==1 and
if KAAPI_REORDER_TASK_EXEC==1 then a local reordering is possible which depends on
the date where a task comes ready because of its inputs are on GPU, in that case the
kernel is immediately launch. Else  if KAAPI_REORDER_TASK_EXEC==0
it its deferred until all previous inserted tasks have been launch on GPU.

If KAAPI_PIPELINE_GPUTASK==0, the this is the original xkblas algorithm that limit the
number of pending+ready task per GPU.

Prepare code to handle multi-hop (or route) to get data back to GPU at high speed.
Add ordering when lid_src is returning: favor first 1/ validity then performance rank,
or 2/ performance rank then validity